Move saving to sql function to a model function
It's easier to manage the functionality if it is separated out
